+++ packages/keyring/src/index.spec.ts
@@ -27,7 +27,7 @@ describe('keypair', () => {
 
     it('adds the pair', () => {
       expect(
-        keypair.addFromSeed(seedTwo, {}).publicKey()
+        keypair.addFromSeed(seedTwo, {}).publicKey
       ).toEqual(publicKeyTwo);
     });
 
@@ -35,7 +35,7 @@ describe('keypair', () => {
       expect(
         keypair.addFromUri(
           'seed sock milk update focus rotate barely fade car face mechanic mercy'
-        ).address()
+        ).address
       ).toEqual('5DkQP32jP4DVJLWWBRBoZF2tpWjqFrcrTBo6H5NcSk7MxKCC');
     });
 
@@ -43,7 +43,7 @@ describe('keypair', () => {
       setAddressPrefix(68);
 
       expect(
-        keypair.addFromMnemonic('moral movie very draw assault whisper awful rebuild speed purity repeat card', {}).address()
+        keypair.addFromMnemonic('moral movie very draw assault whisper awful rebuild speed purity repeat card', {}).address
       ).toEqual('7sPsxWPE5DzAyPT3VuoJYw5NTGscx9QYN9oddQx4kALKC3hH');
     });
 
@@ -57,7 +57,7 @@ describe('keypair', () => {
 
     it('allows retrieval of a specific item', () => {
       expect(
-        keypair.getPair(publicKeyOne).publicKey()
+        keypair.getPair(publicKeyOne).publicKey
       ).toEqual(publicKeyOne);
     });
 
@@ -65,7 +65,7 @@ describe('keypair', () => {
       expect(
         keypair.addFromJson(
           JSON.parse('{"address":"5GoKvZWG5ZPYL1WUovuHW3zJBWBP5eT8CbqjdRY4Q6iMaQua","encoded":"0xb4a14995d25ab609f3686e9fa45f1fb237cd833f33f00d4b12c51858ca070d96972e47d73aae5eeb0fc06f923826cf0943fdb02c2c2ee30ef52a7912663053940d1da4da66b3a3f520ae07422c1c94b2d95690fca9d1f4a997623bb2923a8833280e19e7f72c3c5cfa343974e60e2b3dc53b404fdaf330756daad5e4e3","encoding":{"content":"pkcs8","type":"xsalsa20-poly1305","version":"0"},"meta":{"isTesting":true,"name":"alice"}}')
-        ).publicKey()
+        ).publicKey
       ).toEqual(
         new Uint8Array([209, 114, 167, 76, 218, 76, 134, 89, 18, 195, 43, 160, 168, 10, 87, 174, 105, 171, 174, 65, 14, 92, 203, 89, 222, 232, 78, 47, 68, 50, 219, 79])
       );
@@ -95,13 +95,13 @@ describe('keypair', () => {
 
     it('creates with dev phrase when only path specified', () => {
       expect(
-        keypair.createFromUri('//Alice').address()
+        keypair.createFromUri('//Alice').address
       ).toEqual('5GrwvaEF5zXb26Fz9rcQpDWS57CtERHpNehXCPcNoHGKutQY');
     });
 
     it('adds the pair', () => {
       expect(
-        keypair.addFromSeed(seedTwo, {}).publicKey()
+        keypair.addFromSeed(seedTwo, {}).publicKey
       ).toEqual(publicKeyTwo);
     });
 
@@ -109,7 +109,7 @@ describe('keypair', () => {
       setAddressPrefix(68);
 
       expect(
-        keypair.addFromMnemonic('moral movie very draw assault whisper awful rebuild speed purity repeat card', {}).address()
+        keypair.addFromMnemonic('moral movie very draw assault whisper awful rebuild speed purity repeat card', {}).address
       ).toEqual('7qQGarA4PWjPPVHG4USn1yuuVZvEHN7XZz8o7EbAp48jayZQ');
     });
 
@@ -123,7 +123,7 @@ describe('keypair', () => {
 
     it('allows retrieval of a specific item', () => {
       expect(
-        keypair.getPair(publicKeyOne).publicKey()
+        keypair.getPair(publicKeyOne).publicKey
       ).toEqual(publicKeyOne);
     });
 
@@ -131,7 +131,7 @@ describe('keypair', () => {
       expect(
         keypair.addFromJson(
           JSON.parse('{"address":"5GoKvZWG5ZPYL1WUovuHW3zJBWBP5eT8CbqjdRY4Q6iMaQua","encoded":"0xb4a14995d25ab609f3686e9fa45f1fb237cd833f33f00d4b12c51858ca070d96972e47d73aae5eeb0fc06f923826cf0943fdb02c2c2ee30ef52a7912663053940d1da4da66b3a3f520ae07422c1c94b2d95690fca9d1f4a997623bb2923a8833280e19e7f72c3c5cfa343974e60e2b3dc53b404fdaf330756daad5e4e3","encoding":{"content":"pkcs8","type":"xsalsa20-poly1305","version":"0"},"meta":{"isTesting":true,"name":"alice"}}')
-        ).publicKey()
+        ).publicKey
       ).toEqual(
         new Uint8Array([209, 114, 167, 76, 218, 76, 134, 89, 18, 195, 43, 160, 168, 10, 87, 174, 105, 171, 174, 65, 14, 92, 203, 89, 222, 232, 78, 47, 68, 50, 219, 79])
       );