+++ packages/keyring/src/keyring.ts
@@ -211,7 +211,7 @@ export default class Keyring implements KeyringInstance {
     return this._pairs
       .all()
       .map(({ publicKey }) =>
-        publicKey()
+        publicKey
       );
   }